refactor(repositories): add repositories in database layer
